% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{.check.compatible_model}
\alias{.check.compatible_model}
\title{Checks compatibility between the model settings and the data object.}
\usage{
.check.compatible_model(model, data)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Checks compatibility between the model settings and the data object.
}
