% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{.fit}
\alias{.fit}
\title{Helper function for model fitting. Fits
without intercept on all variables in the dataset but
the first variables as treatment.}
\usage{
.fit(df)
}
\arguments{
\item{df}{Data frame with at least response and treatment variable}
}
\description{
Helper function for model fitting. Fits
without intercept on all variables in the dataset but
the first variables as treatment.
}
