% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linadjust.anova}
\alias{linadjust.anova}
\title{Perform linear model adjustment using the ANOVA adjustment method}
\usage{
\method{linadjust}{anova}(model, data)
}
\arguments{
\item{model}{Object of class ANOVA}

\item{data}{Object of class RoboDataLinear}
}
\description{
Perform linear model adjustment using the ANOVA adjustment method
}
