% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linadjust}
\alias{linadjust}
\title{Generic function for linear adjustment}
\usage{
linadjust(model, data)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Generic function for linear adjustment
}
