% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linmod.ANCOVA}
\alias{linmod.ANCOVA}
\title{Fits a linear model to use in ANCOVA}
\usage{
\method{linmod}{ANCOVA}(model, data)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Fits a linear model to use in ANCOVA
}
