% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linmod.ANHECOVA}
\alias{linmod.ANHECOVA}
\title{Fits a linear model to use in ANHECOVA}
\usage{
\method{linmod}{ANHECOVA}(model, data)
}
\arguments{
\item{model}{Object of class LinModel}

\item{data}{Object of class RoboDataLinear}
}
\description{
Fits a linear model to use in ANHECOVA
}
