% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{.make.model}
\alias{.make.model}
\title{Makes a model class for the specified adjustment method
with settings for covariate randomization
scheme and vcovHC type.}
\usage{
.make.model(adj_method, car_scheme, vcovHC)
}
\arguments{
\item{adj_method}{Type of adjustment method}

\item{car_scheme}{Type of covariate adaptive randomization}

\item{vcovHC}{Type of heteroskedasticity-consistent SE's}
}
\description{
Makes a model class for the specified adjustment method
with settings for covariate randomization
scheme and vcovHC type.
}
