% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linadjust.ANCOVA}
\alias{linadjust.ANCOVA}
\title{Perform linear model adjustment using the ANCOVA adjustment method}
\usage{
\method{linadjust}{ANCOVA}(model, data)
}
\arguments{
\item{model}{Object of class ANCOVA}

\item{data}{Object of class RoboDataLinear}
}
\description{
Perform linear model adjustment using the ANCOVA adjustment method
}
