% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinAdjust.R
\name{linadjust.ANHECOVA}
\alias{linadjust.ANHECOVA}
\title{Perform linear model adjustment using the ANHECOVA adjustment method}
\usage{
\method{linadjust}{ANHECOVA}(model, data)
}
\arguments{
\item{model}{Object of class ANHECOVA}

\item{data}{Object of class RoboDataLinear}
}
\description{
Perform linear model adjustment using the ANHECOVA adjustment method
}
